requirements.txt
streamlit
google-generativeai
newsapi-python
python-dotenv
requests
beautifulsoup4
